launch: Change launch/sev module to launch module
